BJS2-77986. Подписка на других пользователей
1) Написание класса-валидатора для подписок
2) Проверка в фильтрах на null
