+++ src/main/java/school/faang/user_service/controller/subscriptions/SubscriptionController.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.controller;
+package school.faang.user_service.controller.subscriptions;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,10 +8,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import school.faang.user_service.dto.UserDto;
-import school.faang.user_service.dto.UserFilterDto;
-import school.faang.user_service.exceptions.DataValidationException;
-import school.faang.user_service.service.SubscriptionService;
+import school.faang.user_service.dto.user.UserDto;
+import school.faang.user_service.dto.user.UserFilterDto;
+import school.faang.user_service.service.subscriptions.SubscriptionService;
 
 import java.util.List;
 
@@ -24,19 +23,11 @@ public class SubscriptionController {
 
     @PostMapping("/{followerId}/follow/{followeeId}")
     public void followUser(@PathVariable("followerId") long followerId, @PathVariable("followeeId") long followeeId) {
-        if (checkFollowForSelf(followerId, followeeId)) {
-            throw new DataValidationException("You cant subscribe to yourself");
-        }
-
         subscriptionService.followUser(followerId, followeeId);
     }
 
     @DeleteMapping("/{followerId}/unfollow/{followeeId}")
     public void unfollowUser(@PathVariable("followerId") long followerId, @PathVariable("followeeId") long followeeId) {
-        if (checkFollowForSelf(followerId, followeeId)) {
-            throw new DataValidationException("You cant unsubscribe to yourself");
-        }
-
         subscriptionService.unfollowUser(followerId, followeeId);
     }
 
@@ -60,9 +51,5 @@ public class SubscriptionController {
     @GetMapping("{followerId}/followingCount")
     public Integer getFollowingCount(@PathVariable("followerId") long followerId) {
         return subscriptionService.getFollowingCount(followerId);
-    }
-
-    private boolean checkFollowForSelf(long followerId, long followeeId) {
-        return followerId == followeeId;
     }
 }
