+++ src/main/java/school/faang/user_service/dto/user/UserDto.java
@@ -1,12 +1,17 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.user;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 public class UserDto {
+    @NotNull
     private Long id;
+    @NotBlank
     private String username;
+    @NotBlank
     private String email;
 }
