+++ src/main/java/school/faang/user_service/filters/subscriptions/SubscriptionExperienceFilters.java
@@ -1,7 +1,7 @@
 package school.faang.user_service.filters.subscriptions;
 
 import org.springframework.stereotype.Component;
-import school.faang.user_service.dto.UserFilterDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.entity.User;
 
 import java.time.temporal.ValueRange;
@@ -17,7 +17,7 @@ public class SubscriptionExperienceFilters implements SubscriptionFilter {
 
     @Override
     public boolean apply(User user, UserFilterDto filterDto) {
-        return Objects.nonNull(user.getExperience())  && ValueRange.of(filterDto.getExperienceMin(),
-                        filterDto.getExperienceMax()).isValidIntValue(user.getExperience());
+        return Objects.nonNull(user.getExperience()) && ValueRange.of(filterDto.getExperienceMin(),
+                filterDto.getExperienceMax()).isValidIntValue(user.getExperience());
     }
 }
