+++ src/main/java/school/faang/user_service/filters/subscriptions/SubscriptionFilter.java
@@ -1,6 +1,6 @@
 package school.faang.user_service.filters.subscriptions;
 
-import school.faang.user_service.dto.UserFilterDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.entity.User;
 
 public interface SubscriptionFilter {
