+++ src/main/java/school/faang/user_service/mappers/user/UserMapper.java
@@ -1,7 +1,7 @@
-package school.faang.user_service.mappers;
+package school.faang.user_service.mappers.user;
 
 import org.mapstruct.Mapper;
-import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.entity.User;
 
 @Mapper(componentModel = "spring")
