+++ src/main/java/school/faang/user_service/service/subscriptions/SubscriptionService.java
@@ -1,15 +1,15 @@
-package school.faang.user_service.service;
+package school.faang.user_service.service.subscriptions;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import school.faang.user_service.exceptions.DataValidationException;
-import school.faang.user_service.dto.UserDto;
-import school.faang.user_service.dto.UserFilterDto;
+import school.faang.user_service.dto.user.UserDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.filters.subscriptions.SubscriptionFilter;
-import school.faang.user_service.mappers.UserMapper;
+import school.faang.user_service.mappers.user.UserMapper;
 import school.faang.user_service.repository.SubscriptionRepository;
+import school.faang.user_service.validation.subscriptions.SubscriptionValidations;
 
 import java.util.List;
 import java.util.stream.Stream;
@@ -21,14 +21,13 @@ public class SubscriptionService {
     private final SubscriptionRepository subscriptionRepository;
     private final UserMapper userMapper;
     private final List<SubscriptionFilter> subscriptionFilters;
+    private final SubscriptionValidations subscriptionValidations;
 
     @Transactional
     public void followUser(long followerId, long followeeId) {
         boolean existSub = subscriptionRepository.existsByFollowerIdAndFolloweeId(followerId, followeeId);
-
-        if (existSub) {
-            throw new DataValidationException("You are already subscribed to this user");
-        }
+        subscriptionValidations.ValidateFollowAction(followerId, followeeId);
+        subscriptionValidations.ValidateSubscribeAction(existSub);
 
         subscriptionRepository.followUser(followerId, followeeId);
     }
@@ -36,10 +35,8 @@ public class SubscriptionService {
     @Transactional
     public void unfollowUser(long followerId, long followeeId) {
         boolean existSub = subscriptionRepository.existsByFollowerIdAndFolloweeId(followerId, followeeId);
-
-        if (!existSub) {
-            throw new DataValidationException("You cant unsubscribe to user that you are not subscribed");
-        }
+        subscriptionValidations.ValidateUnfollowAction(followerId, followeeId);
+        subscriptionValidations.ValidateUnsubscribeAction(existSub);
 
         subscriptionRepository.unfollowUser(followerId, followeeId);
     }
